pyproject.toml
[tool.poetry]
name = "smugmug-vision"
version = "0.1.0"
description = "Augment SmugMug photo hosting service with computer vision"
authors = ["David Fenster <david@dfenster.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
rauth = "^0.7.3"
fastai = "^2.7.15"
yarl = "^1.9.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
